LRCI-2533 Move the WorkspaceBranchInformation into TopLevelBuild objects
