+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/TopLevelBuild.java
@@ -30,6 +30,7 @@ import com.liferay.jenkins.results.parser.failure.message.generator.PoshiTestFai
 import com.liferay.jenkins.results.parser.failure.message.generator.PoshiValidationFailureMessageGenerator;
 import com.liferay.jenkins.results.parser.failure.message.generator.RebaseFailureMessageGenerator;
 
+import java.io.File;
 import java.io.IOException;
 import java.io.StringWriter;
 
@@ -655,6 +656,97 @@ public abstract class TopLevelBuild extends BaseBuild {
 						"build_slave_usage_gauge", -1, getMetricLabels()));
 			}
 		}
+	}
+
+	public static class WorkspaceBranchInformation
+		implements BranchInformation {
+
+		@Override
+		public String getCachedRemoteGitRefName() {
+			return _workspaceGitRepository.getGitHubDevBranchName();
+		}
+
+		@Override
+		public String getOriginName() {
+			return _workspaceGitRepository.getSenderBranchUsername();
+		}
+
+		@Override
+		public Integer getPullRequestNumber() {
+			Matcher matcher = _pattern.matcher(
+				_workspaceGitRepository.getGitHubURL());
+
+			if (!matcher.find()) {
+				return 0;
+			}
+
+			return Integer.parseInt(matcher.group("pullNumber"));
+		}
+
+		@Override
+		public String getReceiverUsername() {
+			Matcher matcher = _pattern.matcher(
+				_workspaceGitRepository.getGitHubURL());
+
+			if (!matcher.find()) {
+				return "liferay";
+			}
+
+			return matcher.group("username");
+		}
+
+		@Override
+		public String getRepositoryName() {
+			return _workspaceGitRepository.getName();
+		}
+
+		@Override
+		public String getSenderBranchName() {
+			return _workspaceGitRepository.getSenderBranchName();
+		}
+
+		@Override
+		public String getSenderBranchSHA() {
+			return _workspaceGitRepository.getSenderBranchSHA();
+		}
+
+		@Override
+		public RemoteGitRef getSenderRemoteGitRef() {
+			String remoteURL = JenkinsResultsParserUtil.combine(
+				"git@github.com:", getSenderUsername(), "/",
+				getRepositoryName(), ".git");
+
+			return GitUtil.getRemoteGitRef(
+				getSenderBranchName(), new File("."), remoteURL);
+		}
+
+		@Override
+		public String getSenderUsername() {
+			return _workspaceGitRepository.getSenderBranchUsername();
+		}
+
+		@Override
+		public String getUpstreamBranchName() {
+			return _workspaceGitRepository.getUpstreamBranchName();
+		}
+
+		@Override
+		public String getUpstreamBranchSHA() {
+			return _workspaceGitRepository.getBaseBranchSHA();
+		}
+
+		protected WorkspaceBranchInformation(
+			WorkspaceGitRepository workspaceGitRepository) {
+
+			_workspaceGitRepository = workspaceGitRepository;
+		}
+
+		private static final Pattern _pattern = Pattern.compile(
+			"https://github.com/(?<username>[^/]+)/[^/]/pull/" +
+				"(?<pullNumber>\\d+)");
+
+		private final WorkspaceGitRepository _workspaceGitRepository;
+
 	}
 
 	protected TopLevelBuild(String url) {
